Breakout MacroCompiler into separate classes
