+++ src/main/scala/barstools/macros/MacroCompilerException.scala
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+//TODO: Which license header to use
+
+package barstools.macros
+
+case class MacroCompilerException(msg: String) extends Exception(msg)
